% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ber.R
\name{ber_pal}
\alias{ber_pal}
\title{Bureau for Economic Research color palette (discrete)}
\usage{
ber_pal(palette = "core")
}
\arguments{
\item{palette}{\code{character} The color palette to use. One of the following: \code{core}, \code{darks}, \code{blue_and_gold}, \code{expanded_core}, or \code{navy_alt}.}
}
\description{
The Bureau for Economic Research uses many different color palettes in its plots. Most of these are captured in this package.
}
\details{
These palettes are designed to work well in various combinations, allowing for flexibility in plotting while maintaining visual coherence. Users can select the palette most appropriate for their data visualization needs.
}
\section{Palettes}{

The following palettes are available:

\describe{
\item{core}{A core set of colors combining blues with contrasting shades: "#002E60", "#FF790F", "#0154C3", "#0F76FF", "#61223B", "#b79962"}
\item{darks}{Darker shades ideal for backgrounds or to emphasize certain data points: "#FF790F", "#61223B", "#002E5F", "#A24800", "#3A1423", "#001C39"}
\item{blue_and_gold}{A blend of blues and golds for a classic, regal appearance: "#0F76FF", "#B79961", "#0154C3", "#0046A2", "#735D35", "#013275"}
\item{expanded_core}{An extended version of the core palette with additional variations: "#FF790F", "#002E60", "#0154C3", "#0F76FF", "#FF8D44", "#275A87", "#2370D2","#3B8FFA"}
\item{navy_alt}{Shades of navy, perfect for a subdued, professional look: "#002E60", "#0D386B", "#1E4D7C", "#316490", "#4B7FAA", "#6DA1C6", "#94C3E3", "#C1E1F5"}
}
}

\seealso{
Other related color utilities in this package.

Other colour ber: 
\code{\link{scale_colour_ber}()}
}
\concept{colour ber}
