% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ber.R
\name{geom_event}
\alias{geom_event}
\title{Add Event Annotations to ggplot2 Plots}
\usage{
geom_event(
  start,
  end,
  label = "",
  label_size = 4,
  label_color = "#002E60",
  fill = "darkgrey",
  vjust = 1.5,
  alpha = 0.05
)
}
\arguments{
\item{start}{Starting date of the event.}

\item{end}{Ending date of the event.}

\item{label}{Label to annotate the event.}
}
\value{
A list containing \code{geom_rect} and \code{annotate} functions to
shade the region of interest and label it respectively.
}
\description{
\code{geom_event} provides a convenient way to highlight a specific period or event
on a ggplot2 plot using a rectangle and label.
}
\examples{
\dontrun{
library(ggplot2)
data <- data.frame(Date = as.Date('2020-01-01') + 0:364, Value = rnorm(365))
ggplot(data, aes(x = Date, y = Value)) +
  geom_line() +
  scale_colour_ber("core") +
  geom_event(start = "2020-02-01", end = "2020-04-28", label = "Summer", fill = "#0154C3") +
  theme_ber()
}

}
\seealso{
\code{\link[ggplot2]{geom_rect}}, \code{\link[ggplot2]{annotate}} for the underlying functions.
}
