% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ber.R
\name{geom_reference}
\alias{geom_reference}
\title{Add Reference Line and Label to ggplot2 Plots}
\usage{
geom_reference(
  xdate,
  yintercept,
  label = "",
  label_color = "#002E60",
  label_size = 3,
  lty = 2,
  hjust = 0.5,
  angle = 90,
  vjust = -1
)
}
\arguments{
\item{xdate}{The date where the label should be.}

\item{yintercept}{Y-coordinate at which the reference line should be plotted.}

\item{label}{Label to annotate the reference line.}

\item{label_color}{Color of the label and reference line. Defaults to "#002E60".}

\item{label_size}{Size of the label text. Defaults to 3.}

\item{lty}{Line type for the reference line. Defaults to 2.}

\item{hjust}{Horizontal justification of the label. Defaults to 0.5.}

\item{angle}{Angle at which the label text is to be displayed. Defaults to 90.}

\item{vjust}{Vertical justification of the label. Defaults to -1.}
}
\value{
A custom ggplot2 layer that adds a reference line at the specified yintercept
and a label at the maximum x (date) value.
}
\description{
\code{geom_reference} provides a convenient way to add a horizontal reference line
to a ggplot2 plot with an associated label. The label is positioned at the maximum
date value on the x-axis.
}
\examples{
\dontrun{
library(ggplot2)
data <- data.frame(Date = as.Date('2020-01-01') + 0:364, Value = rnorm(365))
ggplot(data, aes(x = Date, y = Value)) +
  geom_line() +
  geom_reference(yintercept = 0, label = "Reference")
}

}
\seealso{
\code{\link[ggplot2]{geom_hline}}, \code{\link[ggplot2]{annotate}} for the underlying functions.
}
