% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ber.R
\name{theme_ber}
\alias{theme_ber}
\title{Bureau for Economic Research theme}
\usage{
theme_ber(
  base_size = 12,
  base_family = "sans",
  title_size = 14,
  title_family = "Helvetica",
  subtitle_size = 12,
  caption_size = 12,
  subtitle_color = "#FF790F",
  title_color = "#0154C3"
)
}
\arguments{
\item{base_size}{Numeric value specifying the base size for plot elements. Default is 12.}

\item{base_family}{Character string specifying the font family for base plot elements. Default is "sans".}

\item{title_size}{Numeric value specifying the size of the plot title. Default is 14.}

\item{title_family}{Character string specifying the font family for the plot title. Default is "Helvetica".}

\item{subtitle_size}{Numeric value specifying the size of the plot subtitle. Default is 12.}

\item{caption_size}{Numeric value specifying the size of the plot caption. Default is 12.}

\item{subtitle_color}{Character string specifying the color of the plot subtitle. Default is "#FF790F".}

\item{title_color}{Character string specifying the color of the plot title. Default is "#0154C3".}
}
\value{
A ggplot2 theme.
}
\description{
Theme based on the plots in \emph{Bureau for Economic Research} reporting.
}
\details{
This theme should be used with \code{\link{scale_color_ber}()} and \code{\link{scale_fill_ber}()}.
}
\section{Usage}{

To use \code{theme_ber}, simply add it to your ggplot2 plot chain. For instance:

\if{html}{\out{<div class="sourceCode R">}}\preformatted{ggplot(data, aes(x, y)) + 
  geom_point() + 
  theme_ber()
}\if{html}{\out{</div>}}
}

\examples{
\dontrun{
library(ggplot2)
data(mpg)
ggplot(mpg, aes(displ, hwy)) +
  geom_point() +
  theme_ber(title_size = 16, title_family = "Times", title_color = "#FF0000")
}

}
\references{
\url{https://ber.ac.za}
}
\seealso{
Other related themes and utilities in this package.
}
\concept{themes ber}
